requirements.txt
# flask
# sentence-transformers
# faiss-cpu
# requests
# pymilvus
# datasets
# torch
# transformers
# apscheduler

flask==2.3.3
requests==2.31.0
sentence-transformers==2.2.2
pymilvus==2.3.1
python-dotenv==1.0.0
apscheduler==3.10.4
aiohttp==3.8.5
numpy==1.24.3